feat: update barcode last price when purchasing
This should be handled better by Grocy itself, but it does not appear
to be, so we do it manually. Although this does not affect this
CLI application directly and this value is never displayed to the user
when managing barcodes, it is used as the default price on Grocy's web
frontend when the user inputs a barcode for purchasing.