fix: correctly scale amount when purchasing per quantity unit
Grocy unfortunately does not allow setting a quantity unit when
purchasing, so intead the quantity (and price) need to be scaled to the
stock quantity unit.

It also seems like it is only possible to do that when the purchasing
quantity unit is either the stock QU itself (and so no scaling is
necessaryy) or when it is the default purchasing QU (in which case
Grocy's product details include the conversion factor). In order to
support other arbitrary QUs, we would need to make an additional
request [`GET /objects/quantity_unit_conversions_resolved`], but this
commit avoids that by simply reporting an error to the user.